@@ -6,31 +6,31 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.2.1.RELEASE</version>
+        <version>2.7.18</version>
     </parent>
 
     <groupId>ua.com.foxminded</groupId>
-    <artifactId>task-university-java8-api</artifactId>
+    <artifactId>task-university-java11-api</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <org.h2database.version>1.4.200</org.h2database.version>
-        <mockito.version>2.25.1</mockito.version>
-        <commons.lang3.version>3.8.1</commons.lang3.version>
-        <org.flywaydb.flyway.core.version>5.2.4</org.flywaydb.flyway.core.version>
-        <hibernate.jpamodelgen.version>5.4.7.Final</hibernate.jpamodelgen.version>
-        <swagger.version>2.9.2</swagger.version>
-        <database.rider.version>1.9.1</database.rider.version>
+        <java.version>11</java.version>
+        <org.h2database.version>2.1.214</org.h2database.version>
+        <mockito.version>4.11.0</mockito.version>
+        <commons.lang3.version>3.12.0</commons.lang3.version>
+        <org.flywaydb.flyway.core.version>8.5.13</org.flywaydb.flyway.core.version>
+        <hibernate.jpamodelgen.version>5.6.15.Final</hibernate.jpamodelgen.version>
+        <springdoc.version>1.7.0</springdoc.version>
+        <database.rider.version>1.44.0</database.rider.version>
         <commons.beanutils.version>1.9.4</commons.beanutils.version>
 
-        <org.flywaydb.flyway.plugin.version>5.2.4</org.flywaydb.flyway.plugin.version>
-        <maven.checkstyle.plugin.version>3.1.0</maven.checkstyle.plugin.version>
+        <org.flywaydb.flyway.plugin.version>8.5.13</org.flywaydb.flyway.plugin.version>
+        <maven.checkstyle.plugin.version>3.3.0</maven.checkstyle.plugin.version>
         <maven.surefire.plugin.version>2.22.2</maven.surefire.plugin.version>
-        <properties.maven.plugin.version>1.0.0</properties.maven.plugin.version>
+        <properties.maven.plugin.version>1.1.0</properties.maven.plugin.version>
     </properties>
 
     <dependencies>
