@@ -1,22 +1,18 @@
 package ua.com.foxminded.task.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.time.LocalDate;
-import java.util.List;
+import com.github.database.rider.core.api.dataset.DataSet;
+import com.github.database.rider.core.api.dataset.ExpectedDataSet;
+import com.github.database.rider.junit5.api.DBRider;
 
-import javax.sql.DataSource;
-
-import org.junit.Rule;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.github.database.rider.core.DBUnitRule;
-import com.github.database.rider.core.api.dataset.DataSet;
-import com.github.database.rider.core.api.dataset.ExpectedDataSet;
-import com.github.database.rider.junit5.api.DBRider;
+import java.time.LocalDate;
+import java.util.List;
 
 import ua.com.foxminded.task.domain.Student;
 
@@ -27,12 +23,6 @@ public class StudentRepositoryIntegrationTest {
     @Autowired
     private StudentRepository studentRepository;
 
-    @Autowired
-    private DataSource dataSource;
-
-    @Rule
-    public DBUnitRule dbUnitRule = DBUnitRule.instance(() -> dataSource.getConnection());
-
     @Test
     @DataSet(cleanBefore = true, skipCleaningFor = "flyway_schema_history")
     public void whenRepositoryHasNotRecords_thenReturnEmptyList() {
@@ -41,9 +31,9 @@ public class StudentRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "student/students.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/students.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "student/expectedStudents.yml")
     public void whenRepositoryHasRecords_thenReturnNonEmptyList() {
         List<Student> students = studentRepository.findAll();
@@ -51,9 +41,9 @@ public class StudentRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "student/students.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/students.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenPutAtTableDbStudentObjects_thenGetThisObjectsFindById() {
         Student expectedStudent = getStudent();
         Student student = studentRepository.findById(3).get();
@@ -61,9 +51,9 @@ public class StudentRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "student/students.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/students.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenFindByIdFees_thenStudentReturned() {
         int idFees = 111111113;
         Student studentActually = studentRepository.findByIdFees(idFees);
@@ -80,9 +70,9 @@ public class StudentRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "student/student.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/student.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "student/expectedStudent.yml")
     public void whenUpdateObject_thenExpectUpdatedRecord() {
         Student student = getStudent();
@@ -92,32 +82,32 @@ public class StudentRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "student/studentsWithGroups.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/studentsWithGroups.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenFindByGroupId_thenReturnTwoRecords() {
         List<Student> students = studentRepository.findAllByGroupId(2);
         assertThat(students).isNotNull().isNotEmpty().hasSize(3);
     }
 
     @Test
-    @DataSet(value = "student/studentsWithGroups.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/studentsWithGroups.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenCountByGroupId_thenReturnTwoRecords() {
         long count = studentRepository.countByGroupId(2);
         assertThat(count).isEqualTo(3L);
     }
 
     @Test
-    @DataSet(value = "student/studentsWithGroups.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "student/studentsWithGroups.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenCheckExistStudentByIdAndGroupId_thenReturnTrue() {
         boolean isExists = studentRepository.existsStudentByIdAndGroupId(3, 2);
         assertTrue(isExists);
     }
-    
+
     private Student getStudent() {
         Student student = new Student();
         student.setFirstName("firstname3");
@@ -127,5 +117,4 @@ public class StudentRepositoryIntegrationTest {
         student.setIdFees(111111113);
         return student;
     }
-
 }
