@@ -1,25 +1,22 @@
 package ua.com.foxminded.task.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+import com.github.database.rider.core.api.dataset.DataSet;
+import com.github.database.rider.core.api.dataset.ExpectedDataSet;
+import com.github.database.rider.junit5.api.DBRider;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.sql.DataSource;
 import javax.transaction.Transactional;
-
-import org.junit.Rule;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import com.github.database.rider.core.DBUnitRule;
-import com.github.database.rider.core.api.dataset.DataSet;
-import com.github.database.rider.core.api.dataset.ExpectedDataSet;
-import com.github.database.rider.junit5.api.DBRider;
 
 import ua.com.foxminded.task.domain.TimetableItem;
 import ua.com.foxminded.task.domain.repository.TimetableItemModelRepository;
@@ -30,12 +27,6 @@ public class TimetableItemRepositoryIntegrationTest {
 
     @Autowired
     private TimetableItemRepository timetableItemRepository;
-    
-    @Autowired
-    private DataSource dataSource;
-
-    @Rule
-    public DBUnitRule dbUnitRule = DBUnitRule.instance(() -> dataSource.getConnection());
 
     @Test
     @DataSet(cleanBefore = true, skipCleaningFor = "flyway_schema_history")
@@ -46,9 +37,9 @@ public class TimetableItemRepositoryIntegrationTest {
 
     @Test
     @Transactional
-    @DataSet(value = "timetableItem/timetableItems.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "timetableItem/timetableItems.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "timetableItem/expected-timetableItems.yml")
     public void whenRepositoryHasRecords_thenReturnNonEmptyList() {
         List<TimetableItem> timetableItems = timetableItemRepository.findAll();
@@ -57,18 +48,18 @@ public class TimetableItemRepositoryIntegrationTest {
 
     @Test
     @Transactional
-    @DataSet(value = "timetableItem/timetableItems.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "timetableItem/timetableItems.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenPutAtTableDbObjects_thenGetThisObjectsFindById() {
-        int id = 3;
+        int id = 30;
         TimetableItem expectedTimetableItem = TimetableItemModelRepository.getModel3();
         expectedTimetableItem.getTeacher().setSubjects(new ArrayList<>());
         expectedTimetableItem.setId(id);
         TimetableItem actuallyTimetableItem = timetableItemRepository.getOne(id);
         assertThat(expectedTimetableItem).isEqualToComparingFieldByField(actuallyTimetableItem);
     }
-    
+
     @Test
     @DataSet(cleanBefore = true, skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "timetableItem/expected-timetableItem.yml")
@@ -84,9 +75,9 @@ public class TimetableItemRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "timetableItem/timetableItem.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "timetableItem/timetableItem.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "timetableItem/expected-timetableItem.yml")
     public void whenUpdateObject_thenExpectUpdatedRecord() {
         TimetableItem timetableItem = TimetableItemModelRepository.getModel2();
@@ -99,30 +90,29 @@ public class TimetableItemRepositoryIntegrationTest {
         TimetableItem actuallyTimetableItem = timetableItemRepository.saveAndFlush(timetableItem);
         assertThat(actuallyTimetableItem).isNotNull();
     }
-    
+
     @Test
-    @DataSet(value = "timetableItem/timetableItems.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "timetableItem/timetableItems.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenRepositoryHasDuplicateAuditoryLectureAndDate_thenReturnBoolean() {
         LocalDate date = LocalDate.of(2020, 06, 25);
-        TimetableItem exist = timetableItemRepository.findByAuditoryIdAndLectureIdAndDate(1, 1, date); 
+        TimetableItem exist = timetableItemRepository.findByAuditoryIdAndLectureIdAndDate(1, 1, date);
         assertNotNull(exist);
-        TimetableItem noExist = timetableItemRepository.findByAuditoryIdAndLectureIdAndDate(2, 3, date); 
-        assertNull(noExist);
-    }
-    
-    @Test
-    @DataSet(value = "timetableItem/timetableItems.yml", 
-             cleanBefore = true, 
-             cleanAfter = true,
-             skipCleaningFor = "flyway_schema_history")
-    public void whenRepositoryHasDuplicateTeacherLectureAndDate_thenReturnBoolean() {
-        LocalDate date = LocalDate.of(2020, 06, 25);
-        TimetableItem exist = timetableItemRepository.findByTeacherIdAndLectureIdAndDate(1, 1, date); 
-        assertNotNull(exist);
-        TimetableItem noExist = timetableItemRepository.findByTeacherIdAndLectureIdAndDate(1, 4, date); 
+        TimetableItem noExist = timetableItemRepository.findByAuditoryIdAndLectureIdAndDate(2, 3, date);
         assertNull(noExist);
     }
 
+    @Test
+    @DataSet(value = "timetableItem/timetableItems.yml",
+        cleanBefore = true,
+        cleanAfter = true,
+        skipCleaningFor = "flyway_schema_history")
+    public void whenRepositoryHasDuplicateTeacherLectureAndDate_thenReturnBoolean() {
+        LocalDate date = LocalDate.of(2020, 06, 25);
+        TimetableItem exist = timetableItemRepository.findByTeacherIdAndLectureIdAndDate(1, 1, date);
+        assertNotNull(exist);
+        TimetableItem noExist = timetableItemRepository.findByTeacherIdAndLectureIdAndDate(1, 4, date);
+        assertNull(noExist);
+    }
 }
