@@ -1,7 +1,7 @@
 package ua.com.foxminded.task.domain;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
@@ -25,5 +25,4 @@ public class TeacherTest {
         teacher.removeSubject(subject);
         assertFalse(teacher.getSubjects().contains(subject));
     }
-
 }
