@@ -1,9 +1,5 @@
 package ua.com.foxminded.task.controller.rest;
 
-import java.util.List;
-
-import javax.validation.Valid;
-
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,14 +9,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import java.util.List;
+
+import javax.validation.Valid;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import ua.com.foxminded.task.domain.dto.AuditoryDto;
 import ua.com.foxminded.task.service.AuditoryService;
 
 @RestController("auditoryRestController")
 @RequestMapping("/api")
-@Api(description = "Auditory management System", produces = "application/json", consumes = "application/json")
+@Tag(name = "Auditory", description = "Auditory management System")
 public class AuditoryController {
 
     private Logger logger;
@@ -33,21 +33,21 @@ public class AuditoryController {
     }
 
     @GetMapping(path = "/auditories", produces = "application/json")
-    @ApiOperation(value = "View a list of available auditories")
+    @Operation(description = "View a list of available auditories")
     public List<AuditoryDto> getEntities() {
         logger.debug("getEntities()");
         return auditoryService.findAllDto();
     }
 
     @GetMapping(path = "/auditories/{id}", produces = "application/json")
-    @ApiOperation(value = "View the auditory by id")
+    @Operation(description = "View the auditory by id")
     public AuditoryDto getEntityById(@PathVariable("id") int id) {
         logger.debug("getEntityById()");
         return auditoryService.findByIdDto(id);
     }
 
     @PostMapping(path = "/auditories", produces = "application/json")
-    @ApiOperation(value = "Create (if id=0) or update a auditory")
+    @Operation(description = "Create (if id=0) or update a auditory")
     public AuditoryDto saveEntity(@Valid @RequestBody AuditoryDto auditoryDto) {
         logger.debug("saveEntity()");
         if (auditoryDto.getId() != 0) {
@@ -57,5 +57,4 @@ public class AuditoryController {
         }
         return auditoryDto;
     }
-
 }
