@@ -1,9 +1,5 @@
 package ua.com.foxminded.task.controller.rest;
 
-import java.util.List;
-
-import javax.validation.Valid;
-
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,14 +9,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import java.util.List;
+
+import javax.validation.Valid;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import ua.com.foxminded.task.domain.dto.DepartmentDto;
 import ua.com.foxminded.task.service.DepartmentService;
 
 @RestController("departmentRestController")
 @RequestMapping("/api")
-@Api(description = "Department management System", produces = "application/json", consumes = "application/json")
+@Tag(name = "Department", description = "Department management System")
 public class DepartmentController {
 
     private Logger logger;
@@ -33,21 +33,21 @@ public class DepartmentController {
     }
 
     @GetMapping(path = "/departments", produces = "application/json")
-    @ApiOperation(value = "View a list of available departments")
+    @Operation(description = "View a list of available departments")
     public List<DepartmentDto> getEntities() {
         logger.debug("getEntities()");
         return departmentService.findAllDto();
     }
 
     @GetMapping(path = "/departments/{id}", produces = "application/json")
-    @ApiOperation(value = "View the department by id")
+    @Operation(description = "View the department by id")
     public DepartmentDto getEntityById(@PathVariable("id") int id) {
         logger.debug("getEntityById()");
         return departmentService.findByIdDto(id);
     }
 
     @PostMapping(path = "/departments", produces = "application/json")
-    @ApiOperation(value = "Create (if id=0) or update the department")
+    @Operation(description = "Create (if id=0) or update the department")
     public DepartmentDto saveEntity(@Valid @RequestBody DepartmentDto departmentDto) {
         logger.debug("saveEntity()");
         if (departmentDto.getId() != 0) {
@@ -57,5 +57,4 @@ public class DepartmentController {
         }
         return departmentDto;
     }
-
 }
