@@ -1,9 +1,5 @@
 package ua.com.foxminded.task.controller.rest;
 
-import java.util.List;
-
-import javax.validation.Valid;
-
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,14 +9,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import java.util.List;
+
+import javax.validation.Valid;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import ua.com.foxminded.task.domain.dto.TeacherDto;
 import ua.com.foxminded.task.service.TeacherService;
 
 @RestController("teacherRestController")
 @RequestMapping("/api")
-@Api(description = "Teacher management System", produces = "application/json", consumes = "application/json")
+@Tag(name = "Teacher", description = "Teacher management System")
 public class TeacherController {
 
     private Logger logger;
@@ -33,21 +33,21 @@ public class TeacherController {
     }
 
     @GetMapping(path = "/teachers", produces = "application/json")
-    @ApiOperation(value = "View a list of available teachers")
+    @Operation(description = "View a list of available teachers")
     public List<TeacherDto> getEntities() {
         logger.debug("getEntities()");
         return teacherService.findAllDto();
     }
 
     @GetMapping(path = "/teachers/{id}", produces = "application/json")
-    @ApiOperation(value = "View the teacher by id")
+    @Operation(description = "View the teacher by id")
     public TeacherDto getEntityById(@PathVariable("id") int id) {
         logger.debug("getEntityById()");
         return teacherService.findByIdDto(id);
     }
 
     @PostMapping(path = "/teachers", produces = "application/json")
-    @ApiOperation(value = "Create (if id=0) or update the teacher")
+    @Operation(description = "Create (if id=0) or update the teacher")
     public TeacherDto saveEntity(@Valid @RequestBody TeacherDto teacherDto) {
         logger.debug("saveEntity()");
         if (teacherDto.getId() != 0) {
@@ -57,5 +57,4 @@ public class TeacherController {
         }
         return teacherDto;
     }
-
 }
