@@ -1,9 +1,5 @@
 package ua.com.foxminded.task.controller.rest;
 
-import java.util.List;
-
-import javax.validation.Valid;
-
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,14 +9,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import java.util.List;
+
+import javax.validation.Valid;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import ua.com.foxminded.task.domain.dto.TimetableItemDto;
 import ua.com.foxminded.task.service.TimetableItemService;
 
 @RestController("timetableItemRestController")
 @RequestMapping("/api")
-@Api(description = "Timetable item management System", produces = "application/json", consumes = "application/json")
+@Tag(name = "Timetable", description = "Timetable item management System")
 public class TimetableItemController {
 
     private Logger logger;
@@ -33,21 +33,21 @@ public class TimetableItemController {
     }
 
     @GetMapping(path = "/timetable-items", produces = "application/json")
-    @ApiOperation(value = "View a list of available timetable items")
+    @Operation(description = "View a list of available timetable items")
     public List<TimetableItemDto> getEntities() {
         logger.debug("getEntities()");
         return timetableItemService.findAllDto();
     }
 
     @GetMapping(path = "/timetable-items/{id}", produces = "application/json")
-    @ApiOperation(value = "View the timetable item by id")
+    @Operation(description = "View the timetable item by id")
     public TimetableItemDto getEntityById(@PathVariable("id") int id) {
         logger.debug("getEntityById()");
         return timetableItemService.findByIdDto(id);
     }
 
     @PostMapping(path = "/timetable-items", produces = "application/json")
-    @ApiOperation(value = "Create (if id=0) or update the timetable item")
+    @Operation(description = "Create (if id=0) or update the timetable item")
     public TimetableItemDto saveEntity(@Valid @RequestBody TimetableItemDto timetableItemDto) {
         logger.debug("saveEntity()");
         if (timetableItemDto.getId() != 0) {
@@ -57,5 +57,4 @@ public class TimetableItemController {
         }
         return timetableItemDto;
     }
-
 }
