@@ -3,41 +3,39 @@ package ua.com.foxminded.task.domain.dto;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Length;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import ua.com.foxminded.task.validation.annotation.PropertyValueUnique;
 
-@PropertyValueUnique(message = "Auditory number is already exists!", 
-                     nameProperty = "auditoryNumber",
-                     fieldError = "auditoryNumber")
-@ApiModel(description = "Object with a auditory information")
+@PropertyValueUnique(message = "Auditory number is already exists!",
+    nameProperty = "auditoryNumber",
+    fieldError = "auditoryNumber")
+@Schema(description = "Object with a auditory information")
 public class AuditoryDto {
 
     @Min(value = 0, message = "Id must be more than zero!")
-    @ApiModelProperty(notes = "The database generated auditory id")
+    @Schema(description = "The database generated auditory id")
     private int id;
 
     @NotBlank(message = "Title of auditory can`t be blank!")
-    @Length(max = 20, message = "Maximum length is 20!")
-    @ApiModelProperty(notes = "The unique name of auditory. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length is 20!")
+    @Schema(description = "The unique name of auditory. Max length is 20 character")
     private String auditoryNumber;
 
-    @Length(max = 45, message = "Maximum length of auditory type title group is 45!")
-    @ApiModelProperty(notes = "The auditory type title. Max length is 45 character")
+    @Size(max = 45, message = "Maximum length of auditory type title group is 45!")
+    @Schema(description = "The auditory type title. Max length is 45 character")
     private String auditoryTypeTitle;
 
-    @ApiModelProperty(notes = "The unique number of students group")
+    @Schema(description = "The unique number of students group")
     private int auditoryTypeId;
-    
+
     @NotNull(message = "Year of entry can`t be blank or not numeric!")
-    @ApiModelProperty(notes = "The max capacity students in auditory. Value must be numeric.")
+    @Schema(description = "The max capacity students in auditory. Value must be numeric.")
     private Integer maxCapacity;
-    
-    @Length(max = 200, message = "Maximum length of auditory description is 200!")
-    @ApiModelProperty(notes = "The descrioption of auditory. Max length is 200 character")
+
+    @Size(max = 200, message = "Maximum length of auditory description is 200!")
+    @Schema(description = "The descrioption of auditory. Max length is 200 character")
     private String description;
 
     public int getId() {
@@ -102,47 +100,59 @@ public class AuditoryDto {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj)
+        if (this == obj) {
             return true;
-        if (obj == null)
+        }
+        if (obj == null) {
             return false;
-        if (getClass() != obj.getClass())
+        }
+        if (getClass() != obj.getClass()) {
             return false;
+        }
         AuditoryDto other = (AuditoryDto) obj;
         if (auditoryNumber == null) {
-            if (other.auditoryNumber != null)
+            if (other.auditoryNumber != null) {
                 return false;
-        } else if (!auditoryNumber.equals(other.auditoryNumber))
+            }
+        } else if (!auditoryNumber.equals(other.auditoryNumber)) {
             return false;
+        }
         if (auditoryTypeTitle == null) {
-            if (other.auditoryTypeTitle != null)
+            if (other.auditoryTypeTitle != null) {
                 return false;
-        } else if (!auditoryTypeTitle.equals(other.auditoryTypeTitle))
+            }
+        } else if (!auditoryTypeTitle.equals(other.auditoryTypeTitle)) {
             return false;
+        }
         if (description == null) {
-            if (other.description != null)
+            if (other.description != null) {
                 return false;
-        } else if (!description.equals(other.description))
+            }
+        } else if (!description.equals(other.description)) {
             return false;
-        if (auditoryTypeId != other.auditoryTypeId)
+        }
+        if (auditoryTypeId != other.auditoryTypeId) {
             return false;
+        }
         if (maxCapacity == null) {
-            if (other.maxCapacity != null)
+            if (other.maxCapacity != null) {
                 return false;
-        } else if (!maxCapacity.equals(other.maxCapacity))
+            }
+        } else if (!maxCapacity.equals(other.maxCapacity)) {
             return false;
+        }
         return true;
     }
 
     @Override
     public String toString() {
-        return "AuditoryDto [id=" + id 
-             + ", auditoryNumber=" + auditoryNumber 
-             + ", auditoryTypeTitle=" + auditoryTypeTitle 
-             + ", auditoryTypeId=" + auditoryTypeId 
-             + ", maxCapacity=" + maxCapacity
-             + ", description=" + description 
-             + "]";
+        return "AuditoryDto [id=" + id
+            + ", auditoryNumber=" + auditoryNumber
+            + ", auditoryTypeTitle=" + auditoryTypeTitle
+            + ", auditoryTypeId=" + auditoryTypeId
+            + ", maxCapacity=" + maxCapacity
+            + ", description=" + description
+            + "]";
     }
 
 }
