@@ -2,26 +2,24 @@ package ua.com.foxminded.task.domain.dto;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Length;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import ua.com.foxminded.task.validation.annotation.PropertyValueUnique;
 
-@PropertyValueUnique(message = "Auditory type name is already exists!", 
-                     nameProperty = "type",
-                     fieldError = "type")
-@ApiModel(description = "Object with a audotory type information")
+@PropertyValueUnique(message = "Auditory type name is already exists!",
+    nameProperty = "type",
+    fieldError = "type")
+@Schema(description = "Object with a audotory type information")
 public class AuditoryTypeDto {
 
     @Min(value = 0, message = "Id must be more than zero!")
-    @ApiModelProperty(notes = "The database generated auditory type id")
+    @Schema(description = "The database generated auditory type id")
     private int id;
 
     @NotBlank(message = "Type name can`t be blank!")
-    @Length(max = 45, message = "Maximum length is 45!")
-    @ApiModelProperty(notes = "The unique type name. Max length is 45 character")
+    @Size(max = 45, message = "Maximum length is 45!")
+    @Schema(description = "The unique type name. Max length is 45 character")
     private String type;
 
     public int getId() {
@@ -50,26 +48,31 @@ public class AuditoryTypeDto {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj)
+        if (this == obj) {
             return true;
-        if (obj == null)
+        }
+        if (obj == null) {
             return false;
-        if (getClass() != obj.getClass())
+        }
+        if (getClass() != obj.getClass()) {
             return false;
+        }
         AuditoryTypeDto other = (AuditoryTypeDto) obj;
         if (type == null) {
-            if (other.type != null)
+            if (other.type != null) {
                 return false;
-        } else if (!type.equals(other.type))
+            }
+        } else if (!type.equals(other.type)) {
             return false;
+        }
         return true;
     }
 
     @Override
     public String toString() {
-        return "AuditoryTypeDto [id=" + id 
-             + ", type=" + type 
-             + "]";
+        return "AuditoryTypeDto [id=" + id
+            + ", type=" + type
+            + "]";
     }
-    
+
 }
