@@ -1,39 +1,38 @@
 package ua.com.foxminded.task.domain.dto;
 
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.format.annotation.DateTimeFormat.ISO;
+
 import java.time.LocalTime;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Length;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.format.annotation.DateTimeFormat.ISO;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import ua.com.foxminded.task.validation.annotation.PropertyValueUnique;
 
-@PropertyValueUnique(message = "Lecture number is already exists!", 
-                     nameProperty = "number",
-                     fieldError = "number")
-@ApiModel(description = "Object with a lecture time information")
+@PropertyValueUnique(message = "Lecture number is already exists!",
+    nameProperty = "number",
+    fieldError = "number")
+@Schema(description = "Object with a lecture time information")
 public class LectureDto {
 
     @Min(value = 0, message = "Id must be more than zero!")
-    @ApiModelProperty(notes = "The database generated lecture id")
+    @Schema(description = "The database generated lecture id")
     private int id;
-    
+
     @NotBlank(message = "Number name can`t be blank!")
-    @Length(max = 5, message = "Maximum length is 5!")
-    @ApiModelProperty(notes = "The unique number name of lecture. Max length is 5 character")
+    @Size(max = 5, message = "Maximum length is 5!")
+    @Schema(description = "The unique number name of lecture. Max length is 5 character")
     private String number;
 
     @DateTimeFormat(iso = ISO.TIME)
-    @ApiModelProperty(notes = "The time of lecture beginning. Format must be 09:20:00")
+    @Schema(description = "The time of lecture beginning. Format must be 09:20:00")
     private LocalTime startTime;
 
     @DateTimeFormat(iso = ISO.TIME)
-    @ApiModelProperty(notes = "The time of lecture be finished. Format must be 09:20:00")
+    @Schema(description = "The time of lecture be finished. Format must be 09:20:00")
     private LocalTime endTime;
 
     public int getId() {
@@ -80,38 +79,46 @@ public class LectureDto {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj)
+        if (this == obj) {
             return true;
-        if (obj == null)
+        }
+        if (obj == null) {
             return false;
-        if (getClass() != obj.getClass())
+        }
+        if (getClass() != obj.getClass()) {
             return false;
+        }
         LectureDto other = (LectureDto) obj;
         if (endTime == null) {
-            if (other.endTime != null)
+            if (other.endTime != null) {
                 return false;
-        } else if (!endTime.equals(other.endTime))
+            }
+        } else if (!endTime.equals(other.endTime)) {
             return false;
+        }
         if (number == null) {
-            if (other.number != null)
+            if (other.number != null) {
                 return false;
-        } else if (!number.equals(other.number))
+            }
+        } else if (!number.equals(other.number)) {
             return false;
+        }
         if (startTime == null) {
-            if (other.startTime != null)
+            if (other.startTime != null) {
                 return false;
-        } else if (!startTime.equals(other.startTime))
+            }
+        } else if (!startTime.equals(other.startTime)) {
             return false;
+        }
         return true;
     }
 
     @Override
     public String toString() {
-        return "LectureDto [id=" + id 
-             + ", number=" + number 
-             + ", startTime=" + startTime 
-             + ", endTime=" + endTime 
-             + "]";
+        return "LectureDto [id=" + id
+            + ", number=" + number
+            + ", startTime=" + startTime
+            + ", endTime=" + endTime
+            + "]";
     }
-    
 }
