@@ -2,26 +2,24 @@ package ua.com.foxminded.task.domain.dto;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Length;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import ua.com.foxminded.task.validation.annotation.PropertyValueUnique;
 
-@PropertyValueUnique(message = "Subject title is already exists!", 
-                     nameProperty = "title",
-                     fieldError = "title")
-@ApiModel(description = "Object with a subject information")
+@PropertyValueUnique(message = "Subject title is already exists!",
+    nameProperty = "title",
+    fieldError = "title")
+@Schema(description = "Object with a subject information")
 public class SubjectDto {
 
     @Min(value = 0, message = "Id must be more than zero!")
-    @ApiModelProperty(notes = "The database generated subject id")
+    @Schema(description = "The database generated subject id")
     private int id;
 
     @NotBlank(message = "Title can`t be blank!")
-    @Length(max = 20, message = "Maximum length is 20!")
-    @ApiModelProperty(notes = "The unique name of subject. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length is 20!")
+    @Schema(description = "The unique name of subject. Max length is 20 character")
     private String title;
 
     public int getId() {
@@ -50,23 +48,26 @@ public class SubjectDto {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj)
+        if (this == obj) {
             return true;
-        if (obj == null)
+        }
+        if (obj == null) {
             return false;
-        if (getClass() != obj.getClass())
+        }
+        if (getClass() != obj.getClass()) {
             return false;
+        }
         SubjectDto other = (SubjectDto) obj;
-        if (id != other.id)
+        if (id != other.id) {
             return false;
+        }
         return true;
     }
 
     @Override
     public String toString() {
-        return "SubjectDto [id=" + id 
-             + ", title=" + title 
-             + "]";
+        return "SubjectDto [id=" + id
+            + ", title=" + title
+            + "]";
     }
-
 }
