@@ -1,5 +1,7 @@
 package ua.com.foxminded.task.domain.dto;
 
+import org.springframework.format.annotation.DateTimeFormat;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -8,55 +10,52 @@ import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Past;
+import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Length;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import ua.com.foxminded.task.validation.annotation.PropertyValueUnique;
 
-@PropertyValueUnique(message = "Teacher id fees is already exists!", 
-                     nameProperty = "idFees",
-                     fieldError = "idFees")
-@ApiModel(description = "Object with a teacher and department information")
+@PropertyValueUnique(message = "Teacher id fees is already exists!",
+    nameProperty = "idFees",
+    fieldError = "idFees")
+@Schema(description = "Object with a teacher and department information")
 public class TeacherDto {
 
     @Min(value = 0, message = "Id must be more than zero!")
-    @ApiModelProperty(notes = "The database generated teacher id")
+    @Schema(description = "The database generated teacher id")
     private int id;
 
     @NotBlank(message = "First name can`t be blank!")
-    @Length(max = 20, message = "Maximum length of first name is 20!")
-    @ApiModelProperty(notes = "The teacher first name. Not blank. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length of first name is 20!")
+    @Schema(description = "The teacher first name. Not blank. Max length is 20 character")
     private String firstName;
 
-    @Length(max = 20, message = "Maximum length of middle name is 20!")
-    @ApiModelProperty(notes = "The teacher middle name. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length of middle name is 20!")
+    @Schema(description = "The teacher middle name. Max length is 20 character")
     private String middleName;
 
-    @Length(max = 20, message = "Maximum length of last name is 20!")
-    @ApiModelProperty(notes = "The teacher last name. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length of last name is 20!")
+    @Schema(description = "The teacher last name. Max length is 20 character")
     private String lastName;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     @Past(message = "Birthday date must be in past!")
-    @ApiModelProperty(notes = "The teacher birthday date. Date must be in past")
+    @Schema(description = "The teacher birthday date. Date must be in past")
     private LocalDate birthday;
 
     @Min(value = 100000000, message = "Value is 9 number!")
     @Max(value = 999999999, message = "Value is 9 number!")
-    @ApiModelProperty(notes = "The teacher identification number in an fees authority. Value is 9 number")
+    @Schema(description = "The teacher identification number in an fees authority. Value is 9 number")
     private int idFees;
 
-    @Length(max = 20, message = "Maximum length of title department is 20!")
-    @ApiModelProperty(notes = "The teacher name of department. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length of title department is 20!")
+    @Schema(description = "The teacher name of department. Max length is 20 character")
     private String departmentTitle;
 
-    @ApiModelProperty(notes = "The unique number of teacher department")
+    @Schema(description = "The unique number of teacher department")
     private int departmentId;
-    
-    @ApiModelProperty(notes = "The list of subjects by teacher")
+
+    @Schema(description = "The list of subjects by teacher")
     private List<SubjectDto> subjects = new ArrayList<>();
 
     public int getId() {
@@ -148,62 +147,78 @@ public class TeacherDto {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj)
+        if (this == obj) {
             return true;
-        if (obj == null)
+        }
+        if (obj == null) {
             return false;
-        if (getClass() != obj.getClass())
+        }
+        if (getClass() != obj.getClass()) {
             return false;
+        }
         TeacherDto other = (TeacherDto) obj;
         if (birthday == null) {
-            if (other.birthday != null)
+            if (other.birthday != null) {
                 return false;
-        } else if (!birthday.equals(other.birthday))
+            }
+        } else if (!birthday.equals(other.birthday)) {
             return false;
+        }
         if (departmentTitle == null) {
-            if (other.departmentTitle != null)
+            if (other.departmentTitle != null) {
                 return false;
-        } else if (!departmentTitle.equals(other.departmentTitle))
+            }
+        } else if (!departmentTitle.equals(other.departmentTitle)) {
             return false;
+        }
         if (firstName == null) {
-            if (other.firstName != null)
+            if (other.firstName != null) {
                 return false;
-        } else if (!firstName.equals(other.firstName))
+            }
+        } else if (!firstName.equals(other.firstName)) {
             return false;
-        if (departmentId != other.departmentId)
+        }
+        if (departmentId != other.departmentId) {
             return false;
-        if (idFees != other.idFees)
+        }
+        if (idFees != other.idFees) {
             return false;
+        }
         if (lastName == null) {
-            if (other.lastName != null)
+            if (other.lastName != null) {
                 return false;
-        } else if (!lastName.equals(other.lastName))
+            }
+        } else if (!lastName.equals(other.lastName)) {
             return false;
+        }
         if (middleName == null) {
-            if (other.middleName != null)
+            if (other.middleName != null) {
                 return false;
-        } else if (!middleName.equals(other.middleName))
+            }
+        } else if (!middleName.equals(other.middleName)) {
             return false;
+        }
         if (subjects == null) {
-            if (other.subjects != null)
+            if (other.subjects != null) {
                 return false;
-        } else if (!subjects.equals(other.subjects))
+            }
+        } else if (!subjects.equals(other.subjects)) {
             return false;
+        }
         return true;
     }
 
     @Override
     public String toString() {
-        return "TeacherDto [id=" + id 
-             + ", firstName=" + firstName 
-             + ", middleName=" + middleName 
-             + ", lastName=" + lastName 
-             + ", birthday=" + birthday 
-             + ", idFees=" + idFees 
-             + ", departmentTitle=" + departmentTitle 
-             + ", departmentId=" + departmentId 
-             + ", subjects=" + subjects 
-             + "]";
+        return "TeacherDto [id=" + id
+            + ", firstName=" + firstName
+            + ", middleName=" + middleName
+            + ", lastName=" + lastName
+            + ", birthday=" + birthday
+            + ", idFees=" + idFees
+            + ", departmentTitle=" + departmentTitle
+            + ", departmentId=" + departmentId
+            + ", subjects=" + subjects
+            + "]";
     }
-
 }
