@@ -1,61 +1,60 @@
 package ua.com.foxminded.task.domain.dto;
 
+import org.springframework.format.annotation.DateTimeFormat;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
 import javax.validation.constraints.Min;
+import javax.validation.constraints.Size;
 
-import org.hibernate.validator.constraints.Length;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import ua.com.foxminded.task.validation.annotation.AuditoryAndTimeUnique;
 import ua.com.foxminded.task.validation.annotation.TeacherAndTimeUnique;
 
 @TeacherAndTimeUnique(fieldError = "teacherId", message = "Teacher will be busy at the time!")
 @AuditoryAndTimeUnique(fieldError = "auditoryId", message = "Auditory will be busy at the time!")
-@ApiModel(description = "Object with a full timetableItem information")
+@Schema(description = "Object with a full timetableItem information")
 public class TimetableItemDto {
 
     @Min(value = 0, message = "Id must be more than zero!")
-    @ApiModelProperty(notes = "The database generated timetableItem id")
+    @Schema(description = "The database generated timetableItem id")
     private int id;
 
-    @Length(max = 45, message = "Maximum length of title subject is 45!")
-    @ApiModelProperty(notes = "The subject name. Max length is 45 character")
+    @Size(max = 45, message = "Maximum length of title subject is 45!")
+    @Schema(description = "The subject name. Max length is 45 character")
     private String subjectTitle;
 
-    @ApiModelProperty(notes = "The unique number of subject")
+    @Schema(description = "The unique number of subject")
     private int subjectId;
 
-    @Length(max = 20, message = "Maximum length of title auditory is 20!")
-    @ApiModelProperty(notes = "The auditory title. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length of title auditory is 20!")
+    @Schema(description = "The auditory title. Max length is 20 character")
     private String auditoryTitle;
 
-    @ApiModelProperty(notes = "The unique number of auditory")
+    @Schema(description = "The unique number of auditory")
     private int auditoryId;
 
-    @ApiModelProperty(notes = "The list of groups")
+    @Schema(description = "The list of groups")
     private List<GroupDto> groups = new ArrayList<>();
 
-    @Length(max = 5, message = "Maximum length of title of lecture is 5!")
-    @ApiModelProperty(notes = "The lecture title. Max length is 5 character")
+    @Size(max = 5, message = "Maximum length of title of lecture is 5!")
+    @Schema(description = "The lecture title. Max length is 5 character")
     private String lectureTitle;
 
-    @ApiModelProperty(notes = "The unique number of lecture")
+    @Schema(description = "The unique number of lecture")
     private int lectureId;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd")
-    @ApiModelProperty(notes = "The date of datetable item.")
+    @Schema(description = "The date of datetable item.")
     private LocalDate date;
 
-    @Length(max = 20, message = "Maximum length of teacher first name is 20!")
-    @ApiModelProperty(notes = "The teacher first name. Max length is 20 character")
+    @Size(max = 20, message = "Maximum length of teacher first name is 20!")
+    @Schema(description = "The teacher first name. Max length is 20 character")
     private String teacherTitle;
 
-    @ApiModelProperty(notes = "The unique number of teacher")
+    @Schema(description = "The unique number of teacher")
     private int teacherId;
 
     public int getId() {
@@ -165,68 +164,86 @@ public class TimetableItemDto {
 
     @Override
     public boolean equals(Object obj) {
-        if (this == obj)
+        if (this == obj) {
             return true;
-        if (obj == null)
+        }
+        if (obj == null) {
             return false;
-        if (getClass() != obj.getClass())
+        }
+        if (getClass() != obj.getClass()) {
             return false;
+        }
         TimetableItemDto other = (TimetableItemDto) obj;
-        if (auditoryId != other.auditoryId)
+        if (auditoryId != other.auditoryId) {
             return false;
+        }
         if (auditoryTitle == null) {
-            if (other.auditoryTitle != null)
+            if (other.auditoryTitle != null) {
                 return false;
-        } else if (!auditoryTitle.equals(other.auditoryTitle))
+            }
+        } else if (!auditoryTitle.equals(other.auditoryTitle)) {
             return false;
+        }
         if (date == null) {
-            if (other.date != null)
+            if (other.date != null) {
                 return false;
-        } else if (!date.equals(other.date))
+            }
+        } else if (!date.equals(other.date)) {
             return false;
+        }
         if (groups == null) {
-            if (other.groups != null)
+            if (other.groups != null) {
                 return false;
-        } else if (!groups.equals(other.groups))
+            }
+        } else if (!groups.equals(other.groups)) {
             return false;
-        if (lectureId != other.lectureId)
+        }
+        if (lectureId != other.lectureId) {
             return false;
+        }
         if (lectureTitle == null) {
-            if (other.lectureTitle != null)
+            if (other.lectureTitle != null) {
                 return false;
-        } else if (!lectureTitle.equals(other.lectureTitle))
+            }
+        } else if (!lectureTitle.equals(other.lectureTitle)) {
             return false;
-        if (subjectId != other.subjectId)
+        }
+        if (subjectId != other.subjectId) {
             return false;
+        }
         if (subjectTitle == null) {
-            if (other.subjectTitle != null)
+            if (other.subjectTitle != null) {
                 return false;
-        } else if (!subjectTitle.equals(other.subjectTitle))
+            }
+        } else if (!subjectTitle.equals(other.subjectTitle)) {
             return false;
-        if (teacherId != other.teacherId)
+        }
+        if (teacherId != other.teacherId) {
             return false;
+        }
         if (teacherTitle == null) {
-            if (other.teacherTitle != null)
+            if (other.teacherTitle != null) {
                 return false;
-        } else if (!teacherTitle.equals(other.teacherTitle))
+            }
+        } else if (!teacherTitle.equals(other.teacherTitle)) {
             return false;
+        }
         return true;
     }
 
     @Override
     public String toString() {
-        return "TimetableItemDto [id=" + id 
-             + ", subjectTitle=" + subjectTitle 
-             + ", subjectId=" + subjectId 
-             + ", auditoryTitle=" + auditoryTitle 
-             + ", auditoryId=" + auditoryId 
-             + ", groups=" + groups
-             + ", lectureTitle=" + lectureTitle 
-             + ", lectureId=" + lectureId 
-             + ", date=" + date 
-             + ", teacherTitle=" + teacherTitle 
-             + ", teacherId=" + teacherId 
-             + "]";
+        return "TimetableItemDto [id=" + id
+            + ", subjectTitle=" + subjectTitle
+            + ", subjectId=" + subjectId
+            + ", auditoryTitle=" + auditoryTitle
+            + ", auditoryId=" + auditoryId
+            + ", groups=" + groups
+            + ", lectureTitle=" + lectureTitle
+            + ", lectureId=" + lectureId
+            + ", date=" + date
+            + ", teacherTitle=" + teacherTitle
+            + ", teacherId=" + teacherId
+            + "]";
     }
-    
 }
