@@ -2,19 +2,15 @@ package ua.com.foxminded.task.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.List;
+import com.github.database.rider.core.api.dataset.DataSet;
+import com.github.database.rider.core.api.dataset.ExpectedDataSet;
+import com.github.database.rider.junit5.api.DBRider;
 
-import javax.sql.DataSource;
-
-import org.junit.Rule;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.github.database.rider.core.DBUnitRule;
-import com.github.database.rider.core.api.dataset.DataSet;
-import com.github.database.rider.core.api.dataset.ExpectedDataSet;
-import com.github.database.rider.junit5.api.DBRider;
+import java.util.List;
 
 import ua.com.foxminded.task.domain.AuditoryType;
 import ua.com.foxminded.task.domain.repository.AuditoryTypeModelRepository;
@@ -24,23 +20,19 @@ import ua.com.foxminded.task.domain.repository.AuditoryTypeModelRepository;
 public class AuditoryTypeRepositoryIntegrationTest {
     @Autowired
     private AuditoryTypeRepository auditoryTypeRepository;
-    @Autowired
-    private DataSource dataSource;
-    @Rule
-    public DBUnitRule dbUnitRule = DBUnitRule.instance(() -> dataSource.getConnection());
-    
+
     @Test
-    @DataSet(cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenRepositoryHasNotRecords_thenReturnEmptyList() {
         List<AuditoryType> auditoryTypes = auditoryTypeRepository.findAll();
         assertThat(auditoryTypes).isNotNull().isEmpty();
     }
 
     @Test
-    @DataSet(value = "auditory-type/auditorytypes.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "auditory-type/auditorytypes.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "auditory-type/expected-auditorytypes.yml")
     public void whenRepositoryHasRecords_thenReturnNonEmptyList() {
         List<AuditoryType> auditoryTypes = auditoryTypeRepository.findAll();
@@ -48,9 +40,9 @@ public class AuditoryTypeRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "auditory-type/auditorytypes.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "auditory-type/auditorytypes.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenPutAtTableDbObjects_thenGetThisObjectsFindById() {
         AuditoryType expectedAuditoryType = AuditoryTypeModelRepository.getModel3();
         int id = 3;
@@ -59,9 +51,9 @@ public class AuditoryTypeRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "auditory-type/auditorytypes.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "auditory-type/auditorytypes.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenFindByType_thenAuditoryTypeReturned() {
         String type = "Laboratory";
         AuditoryType auditoryTypeActually = auditoryTypeRepository.findByType(type);
@@ -78,9 +70,9 @@ public class AuditoryTypeRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "auditory-type/auditorytype.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "auditory-type/auditorytype.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "auditory-type/expected-auditorytype.yml")
     public void whenUpdateObject_thenExpectUpdatedRecord() {
         AuditoryType auditoryType = AuditoryTypeModelRepository.getModel2();
