@@ -2,19 +2,15 @@ package ua.com.foxminded.task.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.List;
+import com.github.database.rider.core.api.dataset.DataSet;
+import com.github.database.rider.core.api.dataset.ExpectedDataSet;
+import com.github.database.rider.junit5.api.DBRider;
 
-import javax.sql.DataSource;
-
-import org.junit.Rule;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.github.database.rider.core.DBUnitRule;
-import com.github.database.rider.core.api.dataset.DataSet;
-import com.github.database.rider.core.api.dataset.ExpectedDataSet;
-import com.github.database.rider.junit5.api.DBRider;
+import java.util.List;
 
 import ua.com.foxminded.task.domain.Group;
 
@@ -25,24 +21,18 @@ public class GroupRepositoryIntegrationTest {
     @Autowired
     private GroupRepository groupRepository;
 
-    @Autowired
-    private DataSource dataSource;
-
-    @Rule
-    public DBUnitRule dbUnitRule = DBUnitRule.instance(() -> dataSource.getConnection());
-
     @Test
     @DataSet(cleanBefore = true,
-             skipCleaningFor = "flyway_schema_history")
+        skipCleaningFor = "flyway_schema_history")
     public void whenRepositoryHasNotRecords_thenReturnEmptyList() {
         List<Group> groups = groupRepository.findAll();
         assertThat(groups).isNotNull().isEmpty();
     }
 
     @Test
-    @DataSet(value = "group/groups.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "group/groups.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "group/expectedGroups.yml")
     public void whenRepositoryHasRecords_thenReturnNonEmptyList() {
         List<Group> groups = groupRepository.findAll();
@@ -50,9 +40,9 @@ public class GroupRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "group/groups.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "group/groups.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenPutAtTableDbGroupObjects_thenGetThisObjectsFindById() {
         Group expectedGroup = new Group();
         expectedGroup.setTitle("group3");
@@ -62,9 +52,9 @@ public class GroupRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "group/groups.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "group/groups.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenFindByTitle_thenGroupReturned() {
         String title = "group2";
         Group groupActually = groupRepository.findByTitle(title);
@@ -81,9 +71,9 @@ public class GroupRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "group/group.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "group/group.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "group/expectedGroup.yml")
     public void whenUpdateObject_thenExpectUpdatedRecord() {
         Group group = getGroup();
@@ -98,5 +88,4 @@ public class GroupRepositoryIntegrationTest {
         group.setYearEntry(2015);
         return group;
     }
-
 }
