@@ -2,21 +2,18 @@ package ua.com.foxminded.task.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.ArrayList;
-import java.util.List;
+import com.github.database.rider.core.api.dataset.DataSet;
+import com.github.database.rider.core.api.dataset.ExpectedDataSet;
+import com.github.database.rider.junit5.api.DBRider;
 
-import javax.sql.DataSource;
-import javax.transaction.Transactional;
-
-import org.junit.Rule;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.github.database.rider.core.DBUnitRule;
-import com.github.database.rider.core.api.dataset.DataSet;
-import com.github.database.rider.core.api.dataset.ExpectedDataSet;
-import com.github.database.rider.junit5.api.DBRider;
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.transaction.Transactional;
 
 import ua.com.foxminded.task.domain.Teacher;
 import ua.com.foxminded.task.domain.repository.TeacherModelRepository;
@@ -27,12 +24,6 @@ public class TeacherRepositoryIntegrationTest {
 
     @Autowired
     private TeacherRepository teacherRepository;
-    
-    @Autowired
-    private DataSource dataSource;
-
-    @Rule
-    public DBUnitRule dbUnitRule = DBUnitRule.instance(() -> dataSource.getConnection());
 
     @Test
     @DataSet(cleanBefore = true, skipCleaningFor = "flyway_schema_history")
@@ -42,9 +33,9 @@ public class TeacherRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "teacher/teachers.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "teacher/teachers.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "teacher/expected-teachers.yml")
     public void whenRepositoryHasRecords_thenReturnNonEmptyList() {
         List<Teacher> teachers = teacherRepository.findAll();
@@ -53,9 +44,9 @@ public class TeacherRepositoryIntegrationTest {
 
     @Test
     @Transactional
-    @DataSet(value = "teacher/teachers.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "teacher/teachers.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenPutAtTableDbObjects_thenGetThisObjectsFindById() {
         Teacher expectedTeacher = TeacherModelRepository.getModel3();
         expectedTeacher.setDepartment(null);
@@ -67,15 +58,15 @@ public class TeacherRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "teacher/teachers.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "teacher/teachers.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     public void whenFindByIdFees_thenTeacherReturned() {
         int idFees = 333111111;
         Teacher teacherActually = teacherRepository.findByIdFees(idFees);
         assertThat(teacherActually.getIdFees()).isEqualTo(idFees);
     }
-    
+
     @Test
     @DataSet(cleanBefore = true, skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "teacher/expected-teacher.yml")
@@ -88,9 +79,9 @@ public class TeacherRepositoryIntegrationTest {
     }
 
     @Test
-    @DataSet(value = "teacher/teacher.yml", 
-             cleanBefore = true, 
-             skipCleaningFor = "flyway_schema_history")
+    @DataSet(value = "teacher/teacher.yml",
+        cleanBefore = true,
+        skipCleaningFor = "flyway_schema_history")
     @ExpectedDataSet(value = "teacher/expected-teacher.yml")
     public void whenUpdateObject_thenExpectUpdatedRecord() {
         Teacher teacher = TeacherModelRepository.getModel2();
@@ -100,32 +91,4 @@ public class TeacherRepositoryIntegrationTest {
         Teacher studentActually = teacherRepository.saveAndFlush(teacher);
         assertThat(studentActually).isNotNull();
     }
-
-//    @Test
-//    @DataSet(value = "student/studentsWithGroups.yml", 
-//             cleanBefore = true, 
-//             skipCleaningFor = "flyway_schema_history")
-//    public void whenFindByGroupId_thenReturnTwoRecords() {
-//        List<Teacher> students = teacherRepository.findAllByGroupId(2);
-//        assertThat(students).isNotNull().isNotEmpty().hasSize(3);
-//    }
-
-//    @Test
-//    @DataSet(value = "student/studentsWithGroups.yml", 
-//             cleanBefore = true, 
-//             skipCleaningFor = "flyway_schema_history")
-//    public void whenCountByGroupId_thenReturnTwoRecords() {
-//        long count = teacherRepository.countByGroupId(2);
-//        assertThat(count).isEqualTo(3L);
-//    }
-
-//    @Test
-//    @DataSet(value = "student/studentsWithGroups.yml", 
-//             cleanBefore = true, 
-//             skipCleaningFor = "flyway_schema_history")
-//    public void whenCheckExistStudentByIdAndGroupId_thenReturnTrue() {
-//        boolean isExists = teacherRepository.existsStudentByIdAndGroupId(3, 2);
-//        assertTrue(isExists);
-//    }
-    
 }
