@@ -1,4 +1,5 @@
 package ua.com.foxminded.task.service.impl;
+
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.anyInt;
@@ -7,10 +8,6 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import java.util.List;
-
-import javax.persistence.EntityNotFoundException;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -18,6 +15,10 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.slf4j.Logger;
 import org.springframework.dao.DataIntegrityViolationException;
+
+import java.util.List;
+
+import javax.persistence.EntityNotFoundException;
 
 import ua.com.foxminded.task.dao.GroupRepository;
 import ua.com.foxminded.task.dao.StudentRepository;
@@ -28,7 +29,7 @@ import ua.com.foxminded.task.domain.dto.StudentDto;
 import ua.com.foxminded.task.domain.repository.StudentModelRepository;
 import ua.com.foxminded.task.domain.repository.dto.StudentDtoModelRepository;
 
-public class StudentServiceImplTest {
+class StudentServiceImplTest {
 
     @Mock
     private Logger logger;
@@ -40,8 +41,8 @@ public class StudentServiceImplTest {
     private StudentServiceImpl studentService;
 
     @BeforeEach
-    public void initMocks() {
-        MockitoAnnotations.initMocks(this);
+    void initMocks() {
+        MockitoAnnotations.openMocks(this);
     }
 
     @Test
@@ -100,68 +101,68 @@ public class StudentServiceImplTest {
         verify(groupRepository, times(1)).getOne(student.getGroup().getId());
         assertEquals(studentDto, studentDtoActually);
     }
-    
+
     @Test
-    public void whenFindByIdFees_thenInvokeMethod() {
+    void whenFindByIdFees_thenInvokeMethod() {
         Student expectedStudent = StudentModelRepository.getModel1();
         Integer expectedIdFees = expectedStudent.getIdFees();
         doReturn(expectedStudent).when(studentRepository).findByIdFees(expectedIdFees);
-        
+
         Student actuallyStudent = studentService.findByIdFees(expectedIdFees);
-        
+
         verify(studentRepository, times(1)).findByIdFees(expectedIdFees);
         assertEquals(expectedIdFees, actuallyStudent.getIdFees());
     }
-    
+
     @Test
-    public void whenCountByGroupId_thenInvokeMethod() {
+    void whenCountByGroupId_thenInvokeMethod() {
         Integer id = 1;
         long expectedCount = 5L;
         doReturn(expectedCount).when(studentRepository).countByGroupId(id);
-        
+
         long actuallyCount = studentService.countByGroupId(id);
-        
+
         verify(studentRepository, times(1)).countByGroupId(id);
         assertEquals(expectedCount, actuallyCount);
     }
-    
+
     @Test
-    public void whenCheckExistsStudentByIdAndGroupId_thenInvokeMethod() {
+    void whenCheckExistsStudentByIdAndGroupId_thenInvokeMethod() {
         Integer studentId = 1;
         Integer groupId = 1;
         boolean expectedResult = true;
         doReturn(expectedResult).when(studentRepository).existsStudentByIdAndGroupId(studentId, groupId);
-        
+
         boolean actuallyResult = studentService.existsStudentByIdAndGroupId(studentId, groupId);
-        
+
         verify(studentRepository, times(1)).existsStudentByIdAndGroupId(studentId, groupId);
         assertEquals(expectedResult, actuallyResult);
     }
-    
+
     @Test
-    public void whenCreateRecordEntityWithId_thenThrowException() {
+    void whenCreateRecordEntityWithId_thenThrowException() {
         StudentDto student = new StudentDto();
         student.setId(1);
 
         assertThatThrownBy(() -> studentService.create(student))
-             .isInstanceOf(EntityAlreadyExistsException.class)
-             .hasMessage("create() studentDto: %s", student);
+            .isInstanceOf(EntityAlreadyExistsException.class)
+            .hasMessage("create() studentDto: %s", student);
     }
-    
+
     @Test
-    public void whenCreateRecordWithNotValidEntity_thenThrowException() {
+    void whenCreateRecordWithNotValidEntity_thenThrowException() {
         StudentDto studentDto = StudentDtoModelRepository.getModel1();
         Student student = StudentModelRepository.getModel1();
         doReturn(student.getGroup()).when(groupRepository).getOne(student.getGroup().getId());
         doThrow(DataIntegrityViolationException.class).when(studentRepository).saveAndFlush(student);
 
         assertThatThrownBy(() -> studentService.create(studentDto))
-             .isInstanceOf(EntityNotValidException.class)
-             .hasMessageContaining("create() student: " + student);
+            .isInstanceOf(EntityNotValidException.class)
+            .hasMessageContaining("create() student: " + student);
     }
-    
+
     @Test
-    public void whenUpdateRecordWithNotValidEntity_thenThrowException() {
+    void whenUpdateRecordWithNotValidEntity_thenThrowException() {
         StudentDto studentDto = StudentDtoModelRepository.getModel1();
         Student student = StudentModelRepository.getModel1();
         doReturn(student.getGroup()).when(groupRepository).getOne(student.getGroup().getId());
@@ -169,18 +170,18 @@ public class StudentServiceImplTest {
         doReturn(true).when(studentRepository).existsById(student.getId());
 
         assertThatThrownBy(() -> studentService.update(studentDto))
-             .isInstanceOf(EntityNotValidException.class)
-             .hasMessageContaining("update() student: " + student);
+            .isInstanceOf(EntityNotValidException.class)
+            .hasMessageContaining("update() student: " + student);
     }
-    
+
     @Test
-    public void whenUpdateRecordNotFpund_thenThrowException() {
+    void whenUpdateRecordNotFpund_thenThrowException() {
         StudentDto studentDto = StudentDtoModelRepository.getModel1();
         Student student = StudentModelRepository.getModel1();
         doReturn(false).when(studentRepository).existsById(student.getId());
 
         assertThatThrownBy(() -> studentService.update(studentDto))
-             .isInstanceOf(EntityNotFoundException.class)
-             .hasMessageContaining("Student not exist!");
+            .isInstanceOf(EntityNotFoundException.class)
+            .hasMessageContaining("Student not exist!");
     }
 }
